refactor(ui): put all ui-related files into ui directory
