+++ app/src/main/AndroidManifest.xml
@@ -47,22 +47,19 @@
         </service>
 
         <!-- Settings activity -->
-        <activity android:name=".settings.PrefMainActivity"
+        <activity android:name=".ui.main.PrefMainActivity"
             android:label="@string/trime_app_name"
             android:launchMode="singleTask"/>
 
-        <activity android:name=".settings.AboutActivity"
-            android:label="@string/settings__about_title" />
-
         <activity
-            android:name=".setup.SetupActivity"
+            android:name=".ui.setup.SetupActivity"
             android:exported="false" />
 
         <activity
-            android:name=".settings.LogActivity"
+            android:name=".ui.main.LogActivity"
             android:exported="false"
             android:launchMode="singleTask"
-            android:parentActivityName=".settings.PrefMainActivity" />
+            android:parentActivityName=".ui.main.PrefMainActivity" />
 
         <!-- Using an activity alias to disable/enable the app icon in the launcher -->
         <activity-alias
@@ -70,7 +67,7 @@
             android:label="@string/trime_app_name"
             android:exported="true"
             android:launchMode="singleTask"
-            android:targetActivity=".settings.PrefMainActivity" >
+            android:targetActivity=".ui.main.PrefMainActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.DEFAULT" />
