+++ app/src/main/java/com/osfans/trime/ime/core/Trime.java
@@ -85,11 +85,11 @@ import com.osfans.trime.ime.text.Candidate;
 import com.osfans.trime.ime.text.Composition;
 import com.osfans.trime.ime.text.ScrollView;
 import com.osfans.trime.ime.text.TextInputManager;
-import com.osfans.trime.settings.PrefMainActivity;
-import com.osfans.trime.settings.components.ColorPickerDialog;
-import com.osfans.trime.settings.components.SchemaPickerDialog;
-import com.osfans.trime.settings.components.SoundPickerDialog;
-import com.osfans.trime.settings.components.ThemePickerDialog;
+import com.osfans.trime.ui.components.ColorPickerDialog;
+import com.osfans.trime.ui.components.SchemaPickerDialog;
+import com.osfans.trime.ui.components.SoundPickerDialog;
+import com.osfans.trime.ui.components.ThemePickerDialog;
+import com.osfans.trime.ui.main.PrefMainActivity;
 import com.osfans.trime.util.ShortcutUtils;
 import com.osfans.trime.util.StringUtils;
 import com.osfans.trime.util.ViewUtils;
