+++ app/src/main/java/com/osfans/trime/ui/components/ResetAssetsDialog.kt
@@ -1,4 +1,4 @@
-package com.osfans.trime.settings.components
+package com.osfans.trime.ui.components
 
 import android.content.Context
 import androidx.appcompat.app.AlertDialog
