+++ app/src/main/java/com/osfans/trime/ui/fragments/LooksFragment.kt
@@ -1,4 +1,4 @@
-package com.osfans.trime.settings.fragments
+package com.osfans.trime.ui.fragments
 
 import android.os.Bundle
 import android.view.Menu
@@ -6,8 +6,8 @@ import androidx.core.view.forEach
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.osfans.trime.R
-import com.osfans.trime.settings.components.ColorPickerDialog
-import com.osfans.trime.settings.components.ThemePickerDialog
+import com.osfans.trime.ui.components.ColorPickerDialog
+import com.osfans.trime.ui.components.ThemePickerDialog
 
 class LooksFragment : PreferenceFragmentCompat() {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
