+++ app/src/main/java/com/osfans/trime/ui/setup/SetupViewModel.kt
@@ -1,4 +1,4 @@
-package com.osfans.trime.setup
+package com.osfans.trime.ui.setup
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
