+++ app/src/main/java/com/osfans/trime/util/ShortcutUtils.kt
@@ -18,7 +18,7 @@ import com.blankj.utilcode.util.IntentUtils
 import com.osfans.trime.core.Rime
 import com.osfans.trime.data.AppPrefs
 import com.osfans.trime.ime.core.Trime
-import com.osfans.trime.settings.LogActivity
+import com.osfans.trime.ui.main.LogActivity
 import timber.log.Timber
 import java.text.FieldPosition
 import java.text.SimpleDateFormat
