+++ app/src/main/res/layout/activity_log.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".settings.LogActivity">
+    tools:context=".ui.main.LogActivity">
 
     <include
         android:id="@+id/toolbar"
@@ -14,7 +14,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"/>
 
-    <com.osfans.trime.settings.components.LogView
+    <com.osfans.trime.ui.components.LogView
         android:id="@+id/logView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
