+++ app/src/main/res/layout/fragment_setup.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="setup.SetupFragment">
+    tools:context=".ui.setup.SetupFragment">
 
     <ImageView
         android:id="@+id/app_icon"
